// src/app/core/services/model/functionTypes.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type FunctionTypes = 'hygieneFlush' | 'circulation' | 'thermalDisinfection';

export const FunctionTypes = {
    HygieneFlush: 'hygieneFlush' as FunctionTypes,
    Circulation: 'circulation' as FunctionTypes,
    ThermalDisinfection: 'thermalDisinfection' as FunctionTypes
};